BCC-172 Export to Nucleus One Cannot Handle Large Files Complete
Modified upload methods to use a Stream, instead of a byte[], to remove the file size limitation.

+++ Ademero.NucleusOneDotNetSdk/ApiModel/DocumentUpload.cs
@@ -45,7 +45,7 @@ namespace Ademero.NucleusOneDotNetSdk.ApiModel
         public string OriginalFilepath { get; set; }
 
         [JsonProperty("OriginalFileSize")]
-        public int OriginalFileSize { get; set; }
+        public long OriginalFileSize { get; set; }
 
         [JsonProperty("FieldIDsAndValues")]
         public Dictionary<string, List<string>> FieldIDsAndValues { get; set; }
